{"schema":"decodex/commit/1","summary":"Consolidate SQLite review and cleanup modules","authority":"manual"}
